Изменить принцип фильтрации фотографий
