+++ js/api.js
@@ -2,13 +2,7 @@ import {showDataAlert} from './alerts.js';
 
 const getData = (onSuccess) => {
   fetch('https://32.javascript.htmlacademy.pro/kekstagram/data')
-    .then((response) => {
-      if (response.ok) {
-        return response.json();
-      } else {
-        throw new Error(`${response.status} — ${response.statusText}`);
-      }
-    })
+    .then((response) => checkResponse(response))
     .then((posts) => {
       onSuccess(posts);
     }).catch(() => showDataAlert());
@@ -18,14 +12,16 @@ function sendData(onSuccess, onFail, body) {
   fetch('https://32.javascript.htmlacademy.pro/kekstagram/', {
     method: 'POST',
     body: body
-  }).then((response) => {
-    if (response.ok) {
-      return response.json();
-    } else {
-      throw new Error(`${response.status} — ${response.statusText}`);
-    }
-  }).then(() => onSuccess())
+  }).then((response) => checkResponse(response)).then(() => onSuccess())
     .catch(() => onFail());
+}
+
+function checkResponse(response) {
+  if (response.ok) {
+    return response.json();
+  } else {
+    throw new Error(`${response.status} — ${response.statusText}`);
+  }
 }
 
 export {getData, sendData};
